fix(tests): refactor blueprint common API tests into seperate files and into classes to reduce boilerplate function decorators
