+++ tests/common/api/v1/test_auth.py
@@ -25,220 +25,168 @@ user_login_payload["email"] = user_register_payload["email"]
     API_CLIENT_PARAMS,
     indirect=True,
 )
-async def test_user_register(api_client: AsyncClient) -> None:
-    """Test that we get a 200 response when registering a user."""
-    response = await api_client.post(
-        f"{BASE_ROUTE}/auth/register", json=user_register_payload
-    )
-    assert response.status_code == status.HTTP_200_OK
+class TestAuth:
+    """Test suite for authentication endpoints."""
 
-    assert int(response.json()["id"])
+    async def test_user_register(self, api_client: AsyncClient) -> None:
+        """Test that we get a 200 response when registering a user."""
+        response = await api_client.post(
+            f"{BASE_ROUTE}/auth/register", json=user_register_payload
+        )
+        assert response.status_code == status.HTTP_200_OK
 
+        assert int(response.json()["id"])
 
-@pytest.mark.asyncio(loop_scope="session")
-@pytest.mark.parametrize(
-    "api_client",
-    API_CLIENT_PARAMS,
-    indirect=True,
-)
-async def test_user_register_bad_email(api_client: AsyncClient) -> None:
-    """Test that we get a 422 response when registering a user with an invalid email."""
-    invalid_payload = copy.deepcopy(user_register_payload)
-    invalid_payload["email"] = "invalidemail"
+    async def test_user_register_bad_email(self, api_client: AsyncClient) -> None:
+        """Test that we get a 422 response when registering a user with an invalid email."""
+        invalid_payload = copy.deepcopy(user_register_payload)
+        invalid_payload["email"] = "invalidemail"
 
-    response = await api_client.post(
-        f"{BASE_ROUTE}/auth/register", json=invalid_payload
-    )
-    assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
+        response = await api_client.post(
+            f"{BASE_ROUTE}/auth/register", json=invalid_payload
+        )
+        assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
 
+    async def test_duplicate_user_register(self, api_client: AsyncClient) -> None:
+        """Test that we get a 400 response when registering a user with the same email."""
+        response = await api_client.post(
+            f"{BASE_ROUTE}/auth/register", json=user_register_payload
+        )
+        assert response.status_code == status.HTTP_409_CONFLICT
 
-@pytest.mark.asyncio(loop_scope="session")
-@pytest.mark.parametrize(
-    "api_client",
-    API_CLIENT_PARAMS,
-    indirect=True,
-)
-async def test_duplicate_user_register(api_client: AsyncClient) -> None:
-    """Test that we get a 400 response when registering a user with the same email."""
-    response = await api_client.post(
-        f"{BASE_ROUTE}/auth/register", json=user_register_payload
-    )
-    assert response.status_code == status.HTTP_409_CONFLICT
+        response_json = response.json()
+        assert response_json["detail"] == "User already exists"
 
-    response_json = response.json()
-    assert response_json["detail"] == "User already exists"
+    async def test_duplicate_user_diff_name_pass_register(
+        self,
+        api_client: AsyncClient,
+    ) -> None:
+        """Test that we get a 400 response when registering a user with the same email but a different password and name."""
+        new_user_register_payload = copy.deepcopy(user_register_payload)
+        new_user_register_payload["password"] = "newpassword123"  # noqa: S105 (Testing)
+        new_user_register_payload["name"] = "New Name"
 
+        response = await api_client.post(
+            f"{BASE_ROUTE}/auth/register", json=user_register_payload
+        )
+        assert response.status_code == status.HTTP_409_CONFLICT
 
-@pytest.mark.asyncio(loop_scope="session")
-@pytest.mark.parametrize(
-    "api_client",
-    API_CLIENT_PARAMS,
-    indirect=True,
-)
-async def test_duplicate_user_diff_name_pass_register(api_client: AsyncClient) -> None:
-    """Test that we get a 400 response when registering a user with the same email but a different password and name."""
-    new_user_register_payload = copy.deepcopy(user_register_payload)
-    new_user_register_payload["password"] = "newpassword123"  # noqa: S105 (Testing)
-    new_user_register_payload["name"] = "New Name"
+        response_json = response.json()
+        assert response_json["detail"] == "User already exists"
 
-    response = await api_client.post(
-        f"{BASE_ROUTE}/auth/register", json=user_register_payload
-    )
-    assert response.status_code == status.HTTP_409_CONFLICT
+    async def test_user_register_invalid_payload(self, api_client: AsyncClient) -> None:
+        """Test that we get a 422 response when registering a user with an invalid payload."""
+        invalid_payload = copy.deepcopy(user_register_payload)
+        invalid_payload.pop("email")
 
-    response_json = response.json()
-    assert response_json["detail"] == "User already exists"
+        response = await api_client.post(
+            f"{BASE_ROUTE}/auth/register", json=invalid_payload
+        )
+        assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
 
+        invalid_payload = copy.deepcopy(user_register_payload)
+        invalid_payload.pop("password")
 
-@pytest.mark.asyncio(loop_scope="session")
-@pytest.mark.parametrize(
-    "api_client",
-    API_CLIENT_PARAMS,
-    indirect=True,
-)
-async def test_user_register_invalid_payload(api_client: AsyncClient) -> None:
-    """Test that we get a 422 response when registering a user with an invalid payload."""
-    invalid_payload = copy.deepcopy(user_register_payload)
-    invalid_payload.pop("email")
+        response = await api_client.post(
+            f"{BASE_ROUTE}/auth/register", json=invalid_payload
+        )
+        assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
 
-    response = await api_client.post(
-        f"{BASE_ROUTE}/auth/register", json=invalid_payload
-    )
-    assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
+        invalid_payload = copy.deepcopy(user_register_payload)
+        invalid_payload.pop("name")
 
-    invalid_payload = copy.deepcopy(user_register_payload)
-    invalid_payload.pop("password")
+        response = await api_client.post(
+            f"{BASE_ROUTE}/auth/register", json=invalid_payload
+        )
+        assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
 
-    response = await api_client.post(
-        f"{BASE_ROUTE}/auth/register", json=invalid_payload
-    )
-    assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
+    async def test_user_login_correct_pass(self, api_client: AsyncClient) -> None:
+        """Test that we get a 200 response when logging in a user and get a valid JWT cookie."""
+        response = await api_client.post(
+            f"{BASE_ROUTE}/auth/login", json=user_login_payload
+        )
+        assert response.status_code == status.HTTP_200_OK
 
-    invalid_payload = copy.deepcopy(user_register_payload)
-    invalid_payload.pop("name")
+        # Check that the response contains a success message
+        assert response.json()["success"] is True
 
-    response = await api_client.post(
-        f"{BASE_ROUTE}/auth/register", json=invalid_payload
-    )
-    assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
+        # Check that the cookie is set
+        assert "token" in response.cookies
 
+    async def test_user_login_incorrect_pass(self, api_client: AsyncClient) -> None:
+        """Test that we get a 401 response when logging in a user with an incorrect password."""
+        invalid_payload = copy.deepcopy(user_login_payload)
+        invalid_payload["password"] = "incorrectpassword"  # noqa: S105 (Testing)
 
-@pytest.mark.asyncio(loop_scope="session")
-@pytest.mark.parametrize(
-    "api_client",
-    API_CLIENT_PARAMS,
-    indirect=True,
-)
-async def test_user_login_correct_pass(api_client: AsyncClient) -> None:
-    """Test that we get a 200 response when logging in a user and get a valid JWT cookie."""
-    response = await api_client.post(
-        f"{BASE_ROUTE}/auth/login", json=user_login_payload
-    )
-    assert response.status_code == status.HTTP_200_OK
+        response = await api_client.post(
+            f"{BASE_ROUTE}/auth/login", json=invalid_payload
+        )
+        assert response.status_code == status.HTTP_401_UNAUTHORIZED
 
-    # Check that the response contains a success message
-    assert response.json()["success"] is True
+    async def test_nonexistent_user_login(self, api_client: AsyncClient) -> None:
+        """Test that we get a 401 response when logging in a user that doesn't exist."""
+        invalid_payload = copy.deepcopy(user_login_payload)
+        invalid_payload["email"] = "alex@ufsit.club"
 
-    # Check that the cookie is set
-    assert "token" in response.cookies
+        response = await api_client.post(
+            f"{BASE_ROUTE}/auth/login", json=invalid_payload
+        )
+        assert response.status_code == status.HTTP_401_UNAUTHORIZED
 
+    async def test_user_logout(self, api_client: AsyncClient) -> None:
+        """Test that we can successfully logout a user."""
+        # First login to get the cookie
+        login_response = await api_client.post(
+            f"{BASE_ROUTE}/auth/login", json=user_login_payload
+        )
+        assert login_response.status_code == status.HTTP_200_OK
+        assert "token" in login_response.cookies
 
-@pytest.mark.asyncio(loop_scope="session")
-@pytest.mark.parametrize(
-    "api_client",
-    API_CLIENT_PARAMS,
-    indirect=True,
-)
-async def test_user_login_incorrect_pass(api_client: AsyncClient) -> None:
-    """Test that we get a 401 response when logging in a user with an incorrect password."""
-    invalid_payload = copy.deepcopy(user_login_payload)
-    invalid_payload["password"] = "incorrectpassword"  # noqa: S105 (Testing)
+        # Now logout
+        logout_response = await api_client.post(f"{BASE_ROUTE}/auth/logout")
+        assert logout_response.status_code == status.HTTP_200_OK
 
-    response = await api_client.post(f"{BASE_ROUTE}/auth/login", json=invalid_payload)
-    assert response.status_code == status.HTTP_401_UNAUTHORIZED
+        # Verify the response content
+        assert logout_response.json()["success"] is True
 
+        # Note: We cannot verify cookie deletion in tests because HttpX doesn't support
+        # direct inspection of Set-Cookie headers that clear cookies. In a real browser,
+        # this would clear the cookie.
 
-@pytest.mark.asyncio(loop_scope="session")
-@pytest.mark.parametrize(
-    "api_client",
-    API_CLIENT_PARAMS,
-    indirect=True,
-)
-async def test_nonexistent_user_login(api_client: AsyncClient) -> None:
-    """Test that we get a 401 response when logging in a user that doesn't exist."""
-    invalid_payload = copy.deepcopy(user_login_payload)
-    invalid_payload["email"] = "alex@ufsit.club"
+    async def test_new_user_register_login_logout_flow(
+        self,
+        api_client: AsyncClient,
+    ) -> None:
+        """Test the user flow where a new user registers and then logs in followed by logout."""
+        user_register_payload = copy.deepcopy(base_user_register_payload)
+        user_login_payload = copy.deepcopy(base_user_login_payload)
 
-    response = await api_client.post(f"{BASE_ROUTE}/auth/login", json=invalid_payload)
-    assert response.status_code == status.HTTP_401_UNAUTHORIZED
+        user_register_payload["email"] = f"test-auth-{uuid.uuid4()}@ufsit.club"
+        user_login_payload["email"] = user_register_payload["email"]
 
+        # Register new user
+        response = await api_client.post(
+            f"{BASE_ROUTE}/auth/register", json=user_register_payload
+        )
+        assert response.status_code == status.HTTP_200_OK
 
-@pytest.mark.asyncio(loop_scope="session")
-@pytest.mark.parametrize(
-    "api_client",
-    API_CLIENT_PARAMS,
-    indirect=True,
-)
-async def test_user_logout(api_client: AsyncClient) -> None:
-    """Test that we can successfully logout a user."""
-    # First login to get the cookie
-    login_response = await api_client.post(
-        f"{BASE_ROUTE}/auth/login", json=user_login_payload
-    )
-    assert login_response.status_code == status.HTTP_200_OK
-    assert "token" in login_response.cookies
+        # Capture user ID
+        user_id = int(response.json()["id"])
+        assert user_id
 
-    # Now logout
-    logout_response = await api_client.post(f"{BASE_ROUTE}/auth/logout")
-    assert logout_response.status_code == status.HTTP_200_OK
+        # Login as new user
+        response = await api_client.post(
+            f"{BASE_ROUTE}/auth/login", json=user_login_payload
+        )
+        assert response.status_code == status.HTTP_200_OK
 
-    # Verify the response content
-    assert logout_response.json()["success"] is True
+        # Check that the response contains a success message
+        assert response.json()["success"] is True
 
-    # Note: We cannot verify cookie deletion in tests because HttpX doesn't support
-    # direct inspection of Set-Cookie headers that clear cookies. In a real browser,
-    # this would clear the cookie.
+        # Check that the cookie is set
+        assert "token" in response.cookies
+        assert "enc_key" in response.cookies
 
-
-@pytest.mark.asyncio(loop_scope="session")
-@pytest.mark.parametrize(
-    "api_client",
-    API_CLIENT_PARAMS,
-    indirect=True,
-)
-async def test_new_user_register_login_logout_flow(
-    api_client: AsyncClient,
-) -> None:
-    """Test the user flow where a new user registers and then logs in followed by logout."""
-    user_register_payload = copy.deepcopy(base_user_register_payload)
-    user_login_payload = copy.deepcopy(base_user_login_payload)
-
-    user_register_payload["email"] = f"test-auth-{uuid.uuid4()}@ufsit.club"
-    user_login_payload["email"] = user_register_payload["email"]
-
-    # Register new user
-    response = await api_client.post(
-        f"{BASE_ROUTE}/auth/register", json=user_register_payload
-    )
-    assert response.status_code == status.HTTP_200_OK
-
-    # Capture user ID
-    user_id = int(response.json()["id"])
-    assert user_id
-
-    # Login as new user
-    response = await api_client.post(
-        f"{BASE_ROUTE}/auth/login", json=user_login_payload
-    )
-    assert response.status_code == status.HTTP_200_OK
-
-    # Check that the response contains a success message
-    assert response.json()["success"] is True
-
-    # Check that the cookie is set
-    assert "token" in response.cookies
-    assert "enc_key" in response.cookies
-
-    # Logout
-    response = await api_client.post(f"{BASE_ROUTE}/auth/logout")
-    assert response.status_code == status.HTTP_200_OK
+        # Logout
+        response = await api_client.post(f"{BASE_ROUTE}/auth/logout")
+        assert response.status_code == status.HTTP_200_OK
